fix：react-native-callkeep 添加平台判断
