--- .github/ISSUE_TEMPLATE.md
@@ -1,27 +1,0 @@
-# Bug report
-
-- [ ] I've checked the [example](https://github.com/react-native-webrtc/react-native-callkeep/tree/master/example) to reproduce the issue.
-
-- Reproduced on:
- - [ ] Android
- - [ ] iOS
-
-## Description
-
-
-## Steps to Reproduce
-
-
-## Versions
-    - Callkeep:
-    - React Native:
-    - iOS:
-    - Android:
-    - Phone model: 
-
-
-## Logs
-
-```
-Paste here
-```
