+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+    "CodeCheck.TaskIdsUpdateTime": "2024-08-26 18:33:11",
+    "CodeCheck.TaskIdsForRemoteUrl": "https://github.com/MyAndroidNetZWJ/react-native-callkeep.git",
+    "CodeCheck.TaskIdsForRemoteBranch": "",
+    "CodeCheck.TaskIds": "",
+    "CodeCheck.SubRepoTaskIdsInfo": {},
+    "CodeCheck.AllRelatedRemoteBranch": "",
+    "CodeCheck.RelatedRemoteRepo": "https://github.com/MyAndroidNetZWJ/react-native-callkeep.git",
+    "CodeCheck.RelatedRemoteBranch": ""
+}
